wxo_auth_app/requirements.txt
Flask==3.0.3
Flask-Session==0.8.0
msal==1.31.0
PyJWT==2.9.0
python-dotenv==1.0.1
